Write model-specific run scripts and configuration files

Generates run scripts and configuration files for all analyses specified in the provided settings. Most of the heavy lifting is done by the write.config.* function for your specific ecosystem model (e.g. write.config.ED2, write.config.SIPNET).

Usage

run.write.configs(
  settings,
  ensemble.size,
  input_design,
  write = TRUE,
  posterior.files = rep(NA, length(settings$pfts)),
  overwrite = TRUE
)

Arguments

settings

a PEcAn settings list

ensemble.size

number of ensemble runs

input_design

data frame containing the design matrix describing parameter and input indices, as documented in runModule.run.write.configs().

write

should the runs be written to the database?

posterior.files

Filenames for posteriors for drawing samples for ensemble and sensitivity analysis (e.g. post.distns.Rdata, or prior.distns.Rdata)

overwrite

logical: Replace output files that already exist?

Value

an updated settings list, which includes ensemble IDs for SA and ensemble analysis

Details

The default value for posterior.files is NA, in which case the most recent posterior or prior (in that order) for the workflow is used. When specified, posterior.files should be a vector of filenames with one entry for each PFT. Specify filenames with no path; PFT outdirs will be appended. This forces use of only files within this workflow, to avoid confusion.

Author

David LeBauer, Shawn Serbin, Ryan Kelly, Mike Dietze